Restored compilability in 32-bit mode

+++ Google Toolbox/GTMStackTrace.m
@@ -170,14 +170,14 @@ static NSString *GTMStackTraceFromAddressDescriptors(struct GTMAddressDescriptor
 		}
 		if (descs[i].class_name) {
 			[trace appendFormat:@"#%-2lu %#016lx %s[%s %s]  (%s)",
-			 i, (NSUInteger)descs[i].address,
+			 (unsigned long)i, (unsigned long)(NSUInteger)descs[i].address,
 			 (descs[i].is_class_method ? "+" : "-"),
 			 descs[i].class_name,
 			 (descs[i].symbol ? descs[i].symbol : "??"),
 			 (descs[i].filename ? descs[i].filename : "??")];
 		} else {
 			[trace appendFormat:@"#%-2lu %#016lx %s()  (%s)",
-			 i, (NSUInteger)descs[i].address,
+			 (unsigned long)i, (unsigned long)(NSUInteger)descs[i].address,
 			 (descs[i].symbol ? descs[i].symbol : "??"),
 			 (descs[i].filename ? descs[i].filename : "??")];
 		}
